serverside/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
pyotp==2.9.0
pytz==2023.4
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2023.4
